rename and make unique names
